requirements.txt
numpy<2
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pillow
